Move exec to just-scripts-utils
