+++ packages/just-scripts-utils/src/__tests__/exec.spec.ts
